+++ app/main.py
@@ -1,7 +1,23 @@
+from contextlib import asynccontextmanager
+
 from fastapi import FastAPI
 
-app = FastAPI()
+from config.settings import Settings
+from libs.ynab.ynab_client import YNABClient
 
+
+@asynccontextmanager
+async def lifespan(app: FastAPI):
+    app.state.ynab_client = YNABClient(
+        access_token=Settings.ynab_access_token,
+        async_mode=Settings.ynab_async_mode
+    )
+    try:
+        yield
+    finally:
+        await app.state.ynab_client.aclose()
+
+app = FastAPI(lifespan=lifespan, debug=Settings.debug)
 
 @app.get("/healthcheck")
 async def healthcheck():
